--- a/.gitignore
+++ b/.gitignore
@@ -1,0 +1,2 @@
+conf_pred_unified
+.DS_Store