--- a/.gitignore
+++ b/.gitignore
@@ -1,2 +1,4 @@
 conf_pred_unified
+pred
 .DS_Store
+.env