Optimization using narrow search of polygons and simple multiprocessing. 

[Runtime]
Before : ~20s
narrow search of polygons intersection : ~9s
narrow search of polygons intersection  + prepared polygon(which supports efficient intersection) : ~7s
narrow search of polygons intersection  + prepared polygon(which supports efficient intersection) + multiprocessing : ~3s